pyproject.toml
[project]
name = "rag-search-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "nltk==3.9.1",
    "numpy>=2.4.2",
    "openai>=2.26.0",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.2.3",
]
